# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Конфигурация подключения к базе данных",
        "contact": {}
    },
    "paths": {
        "/links": {
            "get": {
                "description": "Получить все ссылки из базы данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Получить все ссылки",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Link"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает ссылку в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Создание новой ссылки",
                "parameters": [
                    {
                        "description": "Link object",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Link"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Link"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/links/{id}": {
            "get": {
                "description": "Получить ссылку по ее уникальному идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Получить ссылку по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Link"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновить ссылку в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Обновить ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated link object",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Link"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить ссылку из базы данных по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Удалить ссылку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Link": {
            "description": "Модель для работы с ссылками",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "@example \"2025-01-15T00:00:00Z\"",
                    "type": "string"
                },
                "id": {
                    "description": "@example 1",
                    "type": "integer"
                },
                "title": {
                    "description": "@example \"Example\"",
                    "type": "string"
                },
                "url": {
                    "description": "@example \"https://example.com\"",
                    "type": "string"
                }
            }
        }
    }
}